requirements.txt
numpy
pandas
nltk
pdfminer.six
docx
ttkbootstrap
spacy
gender-guesser
pillow
wordcloud
textblob